// src/shared/services/api/models/assessmentHistoryItemDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface AssessmentHistoryItemDto {
  assessment_history_id?: string;
  assessment_id?: string;
  /** @nullable */
  assessment_title?: string | null;
  course_id?: string;
  /** @nullable */
  course_title?: string | null;
  attempted_on?: string;
  score?: number;
  weighted_score?: number;
  passed?: boolean;
}