// src/shared/services/api/models/assessmentSubmitRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { Answer } from './answer';

export interface AssessmentSubmitRequestDto {
  user_id?: string;
  started_on?: string;
  /** @nullable */
  answer?: Answer[] | null;
}