// src/shared/services/api/models/submitAssessmentResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { SubmitAnswerReviewDto } from './submitAnswerReviewDto';
import type { WeakTopicDto } from './weakTopicDto';

export interface SubmitAssessmentResponseDto {
  assessment_history_id?: string;
  assessment_id?: string;
  course_id?: string;
  /** @nullable */
  status?: string | null;
  started_on?: string;
  completed_on?: string;
  duration_taken_minutes?: number;
  total_question?: number;
  correct_answer?: number;
  score?: number;
  weighted_score?: number;
  course_mastery_score?: number;
  passed?: boolean;
  /** @nullable */
  tier_awarded?: string | null;
  /** @nullable */
  weak_topic?: WeakTopicDto[] | null;
  /** @nullable */
  answers?: SubmitAnswerReviewDto[] | null;
}